Moves tests for parse_meeting_date() to separate file
